Refactor App component and implement store for device management; rename Router.vue and adjust z-index values in CSS
