+++ src/assets/base.css
@@ -33,7 +33,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  z-index: 1;
+  z-index: 9999;
   pointer-events: none;
   overflow: hidden;
 }
