+++ src/main.js
@@ -1,6 +1,6 @@
-import { createApp } from 'vue'
-import './assets/main.css'
-import './assets/sakura.js' 
-import App from './App.vue'
+import { createApp } from "vue";
+import "./assets/main.css";
+import "./assets/sakura.js";
+import App from "./App.vue";
 
-createApp(App).mount('#app')
+createApp(App).mount("#app");
