+++ src/pages/SplashScreen.vue
@@ -23,7 +23,7 @@ export default {
   background-size: 100% 400%;
   top: 0;
   left: 0;
-  z-index: 99999 !important;
+  z-index: 99 !important;
 }
