+++ src/store.js
@@ -1,0 +1,8 @@
+import { reactive } from "vue";
+
+export const store = reactive({
+  isAreEnabled: true,
+  disableAr(){
+    this.isAreEnabled = false;
+  }
+});
